Moved pvp status to creature as its based on the creature id.
